First fully-working BCH code.

+++ src/bch.h
@@ -1,5 +1,6 @@
 #ifndef __BCH_H
 #define __BCH_H
+#include <stdlib.h>
 
 struct bch {
 	int m;		// 2^m - 1 is max length, n
@@ -21,17 +22,17 @@ void generate_bch(bch_t *bch, const int *data, int *bb);
 
 int apply_bch(const bch_t *bch, int *recd);
 
-void bytes_to_bits(const int *bytes, int *bit_dest, int num_bits);
+void bytes_to_bits(const uint8_t *bytes, int *bit_dest, int num_bits);
 
-void bits_to_bytes(const int *bits, int *byte_dest, int num_bits);
+void bits_to_bytes(const int *bits, uint8_t *byte_dest, int num_bits);
 
 void swap_format(const int *bits, int *dest, int cutoff, int num_bits);
 
-int rotate_byte(int x);
+uint8_t rotate_byte(uint8_t x);
 
 void rotate_bits(const int *in, int *out, int num_bits);
 
-void print_bytes(const char *msg, const int *bytes, int num_bytes);
+void print_bytes(const char *msg, const uint8_t *bytes, int num_bytes);
 
 void print_bits(const char *msg, const int *bits, int num_bits);
